Fix gpstart register bug for multiple RCU RS CS per thread
The translation script currently uses a single register to capture
the gpstart reads.  This means that this register is overwritten by
a later RCU read-side critical section.  This in turn can obscure the
fact that the earlier critical section illegally spanned a grace period,
which in turn permits invalid executions.  LISA1Rr1G.litmus detects this
misbehavior and this commit fixes that bug.

Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>