Switch initial variable write to the value 3
This commit switches the initial write to u0 from 1 to 3.  This allows
the local write-write (LWW) case to be checked for even if the last
process had a value/data dependency.

This is still a partial solution.  The final write to u0 must change
to the value 1 in order to avoid the "exists" clause depending on
the last process's structure.

Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>