Fix litmus-test syntax error when -ds too early
If a given process has the -ds modifier, which indicates upcoming
dependency ordering, it needs to initialize the next process's r4
register to the initial value of the variable that the first process
stores to.  This is required in case the next process has a control
dependency, in which case it needs to know the initial value of
the variable that it loads from in order to control the branch.

Unfortunately, the current code omits the ";" from the initialization,
which results in syntax errors if it is not the last initialization.
This commit adds the needed ";".

Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>