makelitmustests.sh: Eliminate erroneous litmus-test combinations
The litmus-test generation scripts do not allow a data dependency to
be followed by an address dependency, as this would propagate the
address back two processes, which I am too lazy to deal with.  However,
the current makelitmustests.sh script nevertheless generates such
litmus tests.  This commit therefore changes makelitmustests.sh to
filter out these sorts of litmus tests.

Because the error was terminating the script, this commit also adds
a bunch of litmus tests that were hidden by the resulting error.

Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>