RCULBlitmusgen.awk: Tighten up classification of rf links
Currently, any rf link that is not release-acquire (or better) is marked
"Maybe", but there are a great many other types of rf links that provide
strong guarantees.  This commit tightens up the classification
accordingly, so that any rf link having a release, an assign, or
a full memory barrier is deemed to not allow the cycle.

Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>