RCUlitmusCout.awk: Add casts to writes
This change is intended to permit klitmus7 to hit fewer build errors
when running on automatically generated litmus tests.

Note that plain writes to registers still do not have casts because
herd7 does not like this sort of thing:

$ herd7 -conf linux-kernel.cfg ~/paper/scalability/LWNLinuxMM/litmus/auto/C-LB-Lrw+R-A+R-A+R-Od.litmus
File "/home/paulmck/paper/scalability/LWNLinuxMM/litmus/auto/C-LB-Lrw+R-A+R-A+R-Od.litmus", line 36, characters 3-11: unexpected 'intptr_t' (in prog)

This is not currently a problem because the memory model does not yet
officially support plain accesses.

Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>