manual/kernel: Add tests showing READ_ONCE() affecting outcome
In some cases, adding a READ_ONCE() can affect the outcome of a test,
even if the value returned is not used.  One way this happens is when
the READ_ONCE() introduces a data race.  Another, illustrated by these
tests, is where the added READ_ONCE() causes an smp_rmb() to indirectly
order a preceding write to that same variable.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>