manual/kernel: Add a litmus test with a hidden dependency
This commit adds a litmus test that has a data dependency that can be
hidden by control flow.  In this test, both the taken and the not-taken
branches of an "if" statement must be accounted for in order to properly
analyze the litmus test.  But herd7 looks only at individual executions
in isolation, so fails to see the dependency.

Reviewed-by: Joel Fernandes (Google) <joel@joelfernandes.org>
Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>