manual/oota:  Add Jonas and Alan OOTA examples
Each of these new litmus tests contains the URL of the email message
that I took it from.

Please note that I had to tweak the example leading up to
C-JO-OOTA-4.litmus, and I might well have misinterpreted Jonas's "~"
operator.

Also, C-JO-OOTA-7.litmus includes a "*r2 = a" statement that makes herd7
very unhappy.  On the other hand, initializing registers to the address
of a variable is straight forward, as shown in the resulting litmus test.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>