c2asmfuncs.awk: Create do_bad_input_line() function for error reporting
The addition of "if" statements will result in non-obvious error
conditions, such as misnesting.  This commit therefore pulls the current
error reporting into a separate function so that it can be invoked where
convenient.  (As opposed to working hard to arrange all such errors drop
down into the "none of the above" pattern-free awk block.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>